requirements.txt
streamlit
pandas
psycopg2-binary
python-dotenv
openai
google-genai
scipy
numpy
